Move Intervention struct to new files
Parameter reading is moved into a member function
