+++ Pv_mod/Source.cpp
@@ -37,6 +37,9 @@
 ///     population. It is only called once while the population is        ///
 ///     being initialised.                                                ///
 ///                                                                       ///
+///  -  Intervention.hpp, Intervention.cpp                                ///
+///     The Intervention struct stores intervention parameters.           ///
+///                                                                       ///
 ///                                                                       ///
 ///  The code below is structured as follows:                             ///
 ///                                                                       ///
@@ -60,6 +63,7 @@
 /////////////////////////////////////////////////////////////////////////////
 
 #include "Population.hpp"
+#include "Intervention.hpp"
 
 #include <iostream>
 #include <fstream>
@@ -84,118 +88,6 @@
 //          //                                      //
 //////////////////////////////////////////////////////
 //////////////////////////////////////////////////////
-
-
-/////////////////////////////////////////////////////////////////////////////////////////
-//                                                                                     //
-// 0.4. Define a structure for details of interventions                                //
-//                                                                                     //
-/////////////////////////////////////////////////////////////////////////////////////////
-
-struct Intervention
-{
-    ////////////////////////////////
-    // LLINs
-
-    vector<double> LLIN_year;
-    vector<double> LLIN_cover;
-
-
-    ////////////////////////////////
-    // IRS
-
-    vector<double> IRS_year;
-    vector<double> IRS_cover;
-
-    ////////////////////////////////
-    // First-line treatment - blood-stage drugs
-
-    vector<double> BS_treat_year_on;
-    vector<double> BS_treat_year_off;
-    vector<double> BS_treat_BScover;
-    vector<double> BS_treat_BSeff;
-    vector<double> BS_treat_BSproph;
-
-
-    ////////////////////////////////
-    // First-line treatment - blood-stage drugs
-    // plus primaquine
-
-    vector<double> PQ_treat_year_on;
-    vector<double> PQ_treat_year_off;
-    vector<double> PQ_treat_BScover;
-    vector<double> PQ_treat_BSeff;
-    vector<double> PQ_treat_BSproph;
-    vector<double> PQ_treat_PQavail;
-    vector<double> PQ_treat_PQeff;
-    vector<double> PQ_treat_PQproph;
-    vector<int>    PQ_treat_G6PD_risk;
-    vector<int>    PQ_treat_CYP2D6_risk;
-    vector<int>    PQ_treat_preg_risk;
-    vector<double> PQ_treat_low_age;
-
-
-    ////////////////////////////////
-    // MDA - blood-stage drugs
-
-    vector<double> MDA_BS_year;
-    vector<double> MDA_BS_BScover;
-    vector<double> MDA_BS_BSeff;
-    vector<double> MDA_BS_BSproph;
-
-
-    ////////////////////////////////
-    // MDA - blood-stage drugs + primaquine
-
-    vector<double> MDA_PQ_year;
-    vector<double> MDA_PQ_BScover;
-    vector<double> MDA_PQ_BSeff;
-    vector<double> MDA_PQ_BSproph;
-    vector<double> MDA_PQ_PQavail;
-    vector<double> MDA_PQ_PQeff;
-    vector<double> MDA_PQ_PQproph;
-    vector<int>    MDA_PQ_G6PD_risk;
-    vector<int>    MDA_PQ_CYP2D6_risk;
-    vector<int>    MDA_PQ_preg_risk;
-    vector<double> MDA_PQ_low_age;
-
-
-    ////////////////////////////////
-    // MSAT - RDT or PCR with blood-stage drugs + primaquine
-
-    vector<double> MSAT_PQ_year;
-    vector<int>    MSAT_PQ_RDT_PCR;
-    vector<double> MSAT_PQ_sens;
-    vector<double> MSAT_PQ_spec;
-    vector<double> MSAT_PQ_BScover;
-    vector<double> MSAT_PQ_BSeff;
-    vector<double> MSAT_PQ_BSproph;
-    vector<double> MSAT_PQ_PQavail;
-    vector<double> MSAT_PQ_PQeff;
-    vector<double> MSAT_PQ_PQproph;
-    vector<int>    MSAT_PQ_G6PD_risk;
-    vector<int>    MSAT_PQ_CYP2D6_risk;
-    vector<int>    MSAT_PQ_preg_risk;
-    vector<double> MSAT_PQ_low_age;
-
-
-    ////////////////////////////////
-    // SSAT - blood-stage drugs + primaquine
-
-    vector<double> SSAT_PQ_year;
-    vector<double> SSAT_PQ_sens;
-    vector<double> SSAT_PQ_spec;
-    vector<double> SSAT_PQ_BScover;
-    vector<double> SSAT_PQ_BSeff;
-    vector<double> SSAT_PQ_BSproph;
-    vector<double> SSAT_PQ_PQavail;
-    vector<double> SSAT_PQ_PQeff;
-    vector<double> SSAT_PQ_PQproph;
-    vector<int>    SSAT_PQ_G6PD_risk;
-    vector<int>    SSAT_PQ_CYP2D6_risk;
-    vector<int>    SSAT_PQ_preg_risk;
-    vector<double> SSAT_PQ_low_age;
-};
 
 
 /////////////////////////////////////////////////////////////////////////////////////////
@@ -329,221 +221,8 @@ int main(int argc, char** argv)
     Pv_mod_par.read(parameter_File, mosquito_File);
     PNG_pop.N_pop = Pv_mod_par.N_pop;
 
-    /////////////////////////////////////////////////////////////////////////
-    //                                                                     //
-    // 1.7. Read in intervention coverage                                  //
-    //                                                                     //
-    /////////////////////////////////////////////////////////////////////////
-
-    /////////////////////////////////////////////////////////////////////////
-    // 1.7.1. Read in matrix of years and coverages
-
-    cout << "Read in intervention coverage file............" << endl;
-
-    std::ifstream coverage_Stream(coverage_File);
-
-    if (coverage_Stream.fail())
-    {
-        std::cout << "Failure reading in data." << endl;
-    }
-
-
-    /////////////////////////////////////////////////////////////////////////
-    // 1.7.2. Note that the matrix we read in may have variable size.
-    //        We first read in the first line of the file, and then matching
-    //        the spaces between columns to get the number of interventions 
-    //        rounds.
-    //
-    //        There's very likely a much more effective way to do this.
-
-    string line_read;
-    int N_cov_rounds = 0;
-
-    getline(coverage_Stream, line_read);
-
-
-    std::string str(line_read);
-    std::string str2(" ");
-
-    std::size_t space_find = str.find(str2);
-
-
-    // Go through the string finding spaces (" ")
-
-    while (space_find < 10000)
-    {
-        space_find = line_read.find(str2);
-
-        line_read.erase(0, space_find + 1);
-
-        N_cov_rounds = N_cov_rounds + 1;
-    }
-
-    N_cov_rounds = N_cov_rounds - 2;
-
-    cout << "Number of intervention rounds: " << N_cov_rounds << endl;
-
-    coverage_Stream.close();
-
-
-    /////////////////////////////////////////////////////////////////////////
-    // 1.7.3. Note that the matrix we read in may have variable size.
-
-    std::ifstream coverage_Stream2(coverage_File);
-
-    string discard;
-
-    vector<vector<double>> coverage;
-    coverage.resize(0);
-
-    vector<double> cov_read(N_cov_rounds + 1);
-
-    for (int i = 0; i < 55; i++)
-    {
-        coverage_Stream2 >> discard;
-
-        for (int j = 1; j<(N_cov_rounds + 2); j++)
-        {
-            coverage_Stream2 >> cov_read[j-1];
-        }
-
-        coverage.push_back(cov_read);
-    }
-
-
-    /////////////////////////////////////////////////////////////////////////
-    // 1.7.2. Fill out Intervention structure
-
     Intervention PNG_intven;
-
-    for (int j = 0; j<N_cov_rounds; j++)
-    {
-        //////////////////////////////////////////////////////////////
-        // LLINs
-
-        if ((coverage[0][j] > -0.5) && (coverage[1][j] > -0.5))
-        {
-            PNG_intven.LLIN_year.push_back(  coverage[0][j]*365.0 );
-            PNG_intven.LLIN_cover.push_back( coverage[1][j] );
-        }
-
-
-        //////////////////////////////////////////////////////////////
-        // IRS
-
-        if ((coverage[0][j] > -0.5) && (coverage[2][j] > -0.5))
-        {
-            PNG_intven.IRS_year.push_back(  coverage[0][j]*365.0 );
-            PNG_intven.IRS_cover.push_back( coverage[2][j] );
-        }
-
-
-        //////////////////////////////////////////////////////////////
-        // Front-line treatment - blood-stage drugs
-
-        if ((coverage[0][j] > -0.5) && (coverage[4][j] > -0.5))
-        {
-            PNG_intven.BS_treat_year_on.push_back(  coverage[0][j]*365.0 );
-            PNG_intven.BS_treat_year_off.push_back( coverage[3][j]*365.0 );
-            PNG_intven.BS_treat_BScover.push_back(  coverage[4][j] );
-            PNG_intven.BS_treat_BSeff.push_back(    coverage[5][j] );
-            PNG_intven.BS_treat_BSproph.push_back(  coverage[6][j] );
-        }
-
-
-        //////////////////////////////////////////////////////////////
-        // Front-line treatment - primaquine
-
-        if ((coverage[0][j] > -0.5) && (coverage[8][j] > -0.5))
-        {
-            PNG_intven.PQ_treat_year_on.push_back(     coverage[0][j]*365.0 );
-            PNG_intven.PQ_treat_year_off.push_back(    coverage[7][j]*365.0 );
-            PNG_intven.PQ_treat_BScover.push_back(     coverage[8][j] );
-            PNG_intven.PQ_treat_BSeff.push_back(       coverage[9][j] );
-            PNG_intven.PQ_treat_BSproph.push_back(     coverage[10][j] );
-            PNG_intven.PQ_treat_PQavail.push_back(     coverage[11][j] );
-            PNG_intven.PQ_treat_PQeff.push_back(       coverage[12][j] );
-            PNG_intven.PQ_treat_PQproph.push_back(     coverage[13][j] );
-            PNG_intven.PQ_treat_G6PD_risk.push_back(   (int)(coverage[14][j]) );
-            PNG_intven.PQ_treat_CYP2D6_risk.push_back( (int)(coverage[15][j]) );
-            PNG_intven.PQ_treat_preg_risk.push_back(   (int)(coverage[16][j]) );
-            PNG_intven.PQ_treat_low_age.push_back(     coverage[17][j] );
-        }
-
-
-        //////////////////////////////////////////////////////////////
-        // MDA - blood-stage drugs
-
-        if ((coverage[0][j] > -0.5) && (coverage[18][j] > -0.5))
-        {
-            PNG_intven.MDA_BS_year.push_back(    coverage[0][j]*365.0 );
-            PNG_intven.MDA_BS_BScover.push_back( coverage[18][j] );
-            PNG_intven.MDA_BS_BSeff.push_back(   coverage[19][j] );
-            PNG_intven.MDA_BS_BSproph.push_back( coverage[20][j] );
-        }
-
-
-        //////////////////////////////////////////////////////////////
-        // MDA - blood-stage drugs plus primaquine
-
-        if ((coverage[0][j] > -0.5) && (coverage[21][j] > -0.5))
-        {
-            PNG_intven.MDA_PQ_year.push_back(        coverage[0][j]*365.0 );
-            PNG_intven.MDA_PQ_BScover.push_back(     coverage[21][j] );
-            PNG_intven.MDA_PQ_BSeff.push_back(       coverage[22][j] );
-            PNG_intven.MDA_PQ_BSproph.push_back(     coverage[23][j] );
-            PNG_intven.MDA_PQ_PQavail.push_back(     coverage[24][j] );
-            PNG_intven.MDA_PQ_PQeff.push_back(       coverage[25][j] );
-            PNG_intven.MDA_PQ_PQproph.push_back(     coverage[26][j] );
-            PNG_intven.MDA_PQ_G6PD_risk.push_back(   (int)(coverage[27][j]) );
-            PNG_intven.MDA_PQ_CYP2D6_risk.push_back( (int)(coverage[28][j]) );
-            PNG_intven.MDA_PQ_preg_risk.push_back(   (int)(coverage[29][j]) );
-            PNG_intven.MDA_PQ_low_age.push_back(     coverage[30][j] );
-        }
-
-
-        //////////////////////////////////////////////////////////////
-        // MSAT - blood-stage drugs plus primaquine
-
-        if ((coverage[0][j] > -0.5) && (coverage[31][j] > -0.5))
-        {
-            PNG_intven.MSAT_PQ_year.push_back(        coverage[0][j]*365.0 );
-            PNG_intven.MSAT_PQ_BScover.push_back(     coverage[31][j]);
-            PNG_intven.MSAT_PQ_RDT_PCR.push_back(     coverage[32][j] );
-            PNG_intven.MSAT_PQ_sens.push_back(        coverage[33][j] );
-            PNG_intven.MSAT_PQ_BSeff.push_back(       coverage[34][j] );
-            PNG_intven.MSAT_PQ_BSproph.push_back(     coverage[35][j] );
-            PNG_intven.MSAT_PQ_PQavail.push_back(     coverage[36][j] );
-            PNG_intven.MSAT_PQ_PQeff.push_back(       coverage[37][j] );
-            PNG_intven.MSAT_PQ_PQproph.push_back(     coverage[38][j] );
-            PNG_intven.MSAT_PQ_G6PD_risk.push_back(   (int)(coverage[39][j]) );
-            PNG_intven.MSAT_PQ_CYP2D6_risk.push_back( (int)(coverage[40][j]) );
-            PNG_intven.MSAT_PQ_preg_risk.push_back(   (int)(coverage[41][j]) );
-            PNG_intven.MSAT_PQ_low_age.push_back(     coverage[42][j] );
-        }
-
-        //////////////////////////////////////////////////////////////
-        // SSAT - blood-stage drugs plus primaquine
-
-        if ((coverage[0][j] > -0.5) && (coverage[43][j] > -0.5))
-        {
-            PNG_intven.SSAT_PQ_year.push_back(        coverage[0][j]*365.0 );
-            PNG_intven.SSAT_PQ_BScover.push_back(     coverage[43][j] );
-            PNG_intven.SSAT_PQ_sens.push_back(        coverage[44][j] );
-            PNG_intven.SSAT_PQ_spec.push_back(        coverage[45][j] );
-            PNG_intven.SSAT_PQ_BSeff.push_back(       coverage[46][j] );
-            PNG_intven.SSAT_PQ_BSproph.push_back(     coverage[47][j] );
-            PNG_intven.SSAT_PQ_PQavail.push_back(     coverage[48][j] );
-            PNG_intven.SSAT_PQ_PQeff.push_back(       coverage[49][j] );
-            PNG_intven.SSAT_PQ_PQproph.push_back(     coverage[50][j] );
-            PNG_intven.SSAT_PQ_G6PD_risk.push_back(   (int)(coverage[51][j]) );
-            PNG_intven.SSAT_PQ_CYP2D6_risk.push_back( (int)(coverage[52][j]) );
-            PNG_intven.SSAT_PQ_preg_risk.push_back(   (int)(coverage[53][j]) );
-            PNG_intven.SSAT_PQ_low_age.push_back(     coverage[54][j] );
-        }
-
-    }
-
+    PNG_intven.read(coverage_File);
 
     ///////////////////////////////////////////////////////////////////////////
     //                                                                       //
